finalish update. Removed test/non functioning code/added comments
