+++ src/components/Messages/MessageCard.js
@@ -1,21 +1,24 @@
 //Responsible for generating and displaying chat messages based on database. Created by Brandon W.
 
 import React from 'react';
+import { Button } from "react-bootstrap";
+import './Messages.css'
 //Creates message entries by passing in userId and content of message from database
 const MessageCard = props => {
     let activeUserId = sessionStorage.getItem("activeUserID")
 
+    //One day...
     const editMessage = () => {
-        let messageToEdit = document.querySelector("#message")
-        messageToEdit.value = props.messages.message    
+        window.alert("Feature coming soon")  
     }
     
     return (
         <>
+        <input type="hidden" id="messageId" value="" />
         <div>
-            {props.messages.userId}: {props.messages.message}
+            {props.messages.user}: {props.messages.message}
         {/*checks current UserId against userId of message for displaying of edit functionality*/}
-            {props.messages.userId === parseInt(activeUserId) ? <button type="button" onClick={editMessage}>Edit</button>: ""}
+            {props.messages.userId === parseInt(activeUserId) ? <Button className = "editMessageButton" type="button" variant="custom" onClick={editMessage}>Edit</Button>: ""}
         </div>
         </>
     )
