+++ src/components/Messages/MessageList.js
@@ -1,3 +1,4 @@
+//Responsible for managing message data, constructing new messages, and displaying them to the DOM in real time - Brandon W.
 import React, {useState, useEffect} from 'react';
 import MessageCard  from '../Messages/MessageCard';
 import MessageForm from '../Messages/MessageForm';
@@ -6,16 +7,34 @@ import APIManager from '../../components/Modules/APIManager';
 const MessageList = (props) => {
     const [messages, setMessages] = useState([]);
 
+    const clearInputField = () => {
+        
+        let inputField = document.getElementById("message")
+        inputField.value = ""
+    };
+
+    
     const getMessages = () => {
         return APIManager.GetAll("messages").then(messagesFromAPI => {
             setMessages(messagesFromAPI)
-        });
-    };
+        })
+    }
 
+    const constructNewMessage = message => {
+        if (message !== "") {
+            APIManager.Push("messages", message)
+            .then(()=> APIManager.GetAll("messages")
+            .then(setMessages))
+            clearInputField();
+        } else {
+            window.alert("Please enter a message");
+        }
+    };
+    
     useEffect(() => {
         getMessages();
     }, []);
-
+    
     return (
         <>
         <div>
@@ -26,9 +45,7 @@ const MessageList = (props) => {
                     {...props}
                     />
             )}
-        </div>
-        <div>
-            <MessageForm getMessages={getMessages} />
+            <MessageForm getMessages={getMessages} constructMessage={constructNewMessage} />
         </div>
         </>
     )
