+++ src/components/Messages/Messages.css
@@ -1,0 +1,9 @@
+.postMessageButton.btn.btn-custom {
+    background-color: #e37d64;
+    color: #fff7e7;
+}
+
+.editMessageButton.btn.btn-custom {
+    background-color: #e37d64;
+    color: #fff7e7;
+}
